refactor(api): move Eventhook to api from olink
