+++ goldenmaster/tb_empty_api/shared.py
