+++ goldenmaster/testbed2_api/shared.py
