+++ templates/api/shared.py
